refactor: devided all classes in Application.java into their own java file
docs: README.md style changed with bold effects
