+++ README.md
@@ -3,32 +3,32 @@
 ## 객체 별 구현 기능 목록
 
 ### Application
-- 게임 시스템 그 자체 (main)
+- 게임 메인 시스템 (main)
 
 ### Player
-1. 플레이어 입력 받기 기능 (숫자 추리)
+**1. 플레이어 입력 받기 기능 (숫자 추리)**
 - 예외: 입력이 정수가 아님 / 세자리 숫자가 아님 / Valid 하지 않음 (자릿수 숫자 중복 혹은 0 포함)
 - 입력 예외 상황 시 항상 IllegalArgumentException 던지기
 - 예외 에러 발생시 다시 새로운 입력 가능하게 하는 기능
 
-2. 게임 플레이 -> 결과가 3 스트라이크 일때 까지
+**2. 게임 플레이 -> 결과가 3 스트라이크 일때 까지**
 
-3. 재시작할지 그만둘지 결정
+**3. 재시작할지 그만둘지 결정**
 - 재시작 -> main 재실행 / 그만둠 -> 그냥 패스
 - 예외: 입력이 정수가 아님 / 입력이 1이나 2가 아님
 - 입력 예외 상황 시 항상 IllegalArgumentException 던지기
 - 예외 에러 발생시 다시 새로운 입력 가능하게 하는 기능
 
 ### Computer
-1. 랜덤넘버 생성 (정답)
+**1. 랜덤넘버 생성 (정답)**
 - 예외: 입력이 정수가 아님 / 세자리 숫자가 아님 / Valid 하지 않음 (자릿수 숫자 중복 혹은 0 포함)
 - Valid한 랜덤 넘버 (정답) 가 나올 때 까지 계속 생성하는 기능
 
-2. 플레이어의 숫자 추리 결과를 바탕으로 Ball, Strike 카운트
+**2. 플레이어의 숫자 추리 결과를 바탕으로 Ball, Strike 카운트**
 - 플레이어 입력과 정답을 비교하는 스트라이크와 볼 카운트 기능
 - 정답을 맞출 때 (3 스트라이크) 까지 반복
 
-3. Ball, Strike 카운트 기반으로 결과 파악
+**3. Ball, Strike 카운트 기반으로 결과 파악**
 
 ### StrikeAndBall
 - 스트라이크와 볼 카운트 값을 담는 클래스
