+++ src/main/java/baseball/StrikeAndBall.java
@@ -1,0 +1,20 @@
+package baseball;
+
+// to store Strike and Ball count
+public class StrikeAndBall {
+    int strike;
+    int ball;
+
+    public StrikeAndBall(int strike, int ball) {
+        this.strike = strike;
+        this.ball = ball;
+    }
+
+    public int getStrike() {
+        return strike;
+    }
+
+    public int getBall() {
+        return ball;
+    }
+}
